Revert "conversion error"
This reverts commit f811117458d3747c53be3c7cb8ae85338d26092b.

+++ TombEngine/Scripting/Internal/TEN/Objects/Moveable/MoveableObject.cpp
@@ -525,9 +525,9 @@ void SetLevelFuncCallback(const TypeOrNil<LevelFunc>& cb, const std::string& cal
 
 }
 
-int Moveable::GetIndex() const
+short Moveable::GetIndex() const
 {
-	return (int)m_num;
+	return m_num;
 }
 
 void Moveable::SetOnHit(const TypeOrNil<LevelFunc>& cb)
